Add privacy options for Google Analytics modules
