+++ wrighter-client/constants.ts
@@ -1,1 +1,3 @@
 export const API_BASE_URL = process.env.API_BASE_URL || "http://localhost:8080/api";
+export const PRIVACY = process.env.PRIVACY || false;
+export const GTAG = process.env.GTAG || "G-3GJH7QG23L";
